index_miyatter/requirements.txt
Django==4.0
djangorestframework
django-filter
boto3
django-environ
mecab-python3
